split text messages into a separate module
